+++ test/mysql/brpc_mysql_prepared_integration_unittest.cpp
@@ -52,8 +52,8 @@
 
 #include <brpc/channel.h>
 #include <brpc/controller.h>
-#include <brpc/mysql.h>
-#include <brpc/policy/mysql_authenticator.h>
+#include "brpc/policy/mysql/mysql.h"
+#include "brpc/policy/mysql/mysql_authenticator.h"
 #include "butil/logging.h"
 
 // These flags intentionally mirror the names used by
