+++ test/mysql/brpc_mysql_txn_integration_unittest.cpp
@@ -56,9 +56,9 @@
 #include <string>
 
 #include "brpc/channel.h"
-#include "brpc/mysql.h"
-#include "brpc/mysql_transaction.h"
-#include "brpc/policy/mysql_authenticator.h"
+#include "brpc/policy/mysql/mysql.h"
+#include "brpc/policy/mysql/mysql_transaction.h"
+#include "brpc/policy/mysql/mysql_authenticator.h"
 #include "butil/logging.h"
 
 // These gflags are intentionally re-declared here (not shared with the auth
